Delete the original download command reply
Deletes the bot's reply to the original download command - after the download task finishes/is interrupted - if less than 10 messages have been sent to the group the download started in, since the download started.
This is done because a new status is sent on download finish/interrupt, and the reply to the original download command just clutters the chat.