Use general status instead of individual status
TG starts rate limiting if more than 20 messages are sent (or edited) to a chat per minute. If there are multiple downloads in a group, updating the status messages for each of those downloads easily triggers the rate limit.
This commit gets rid of the individual status messages, replacing them with one single general status message (containing the status of all downloads). This reduces the number of messages that need to be kept up to date.